requirements.txt
fastapi
requests
uvicorn
selenium 
langchain-groq
python-dotenv
pandas
scraper